refactor: 구분자 Enum 클래스 분리
- 기본 구분자와 사용할 수 없는 커스텀 구분자를 각 Enum 클래스로 관리합니다.
